Implemented views.
The window now shows more/less of the screen if resized instead of growing/squashing the contents. You can also move the view left and right with the arrow keys.

It will look wrong, as pressing right will make the tiles go left, but remember you're moving the view not the objects, if you look at an object and walk to the right, it would be the same as you standing still and the object moving left.

I have no idea why I felt the need to explain this.