Implemented FPS:
As we now have movement, we need to keep track of the FPS so we can have smooth movement on all platforms.

This also means I've had to work on getting UI elements to stay put no matter what happens to the views. The current system is based on using the centre and size to go to a specific edge of the view (and thus window), and adding/subtracting a certain amount to put the object in the correct place.