Added score and coins
The coin tile is a tile that, when the player collides with it, gives 100 points and dissapears. Dying now subtracts 50 points, and the score has been added to the UI.

This also means tiles now have handle event functions, which means for now means that tiles that don't have any events to handle have empty functions, which I think will be optimised by the compiler? I don't know.