Added spikes, cleaned up
Spikes are a tile that kills the player on contact. The hitbox need to be reduced slightly but I'll get around to that.

This also prompted a cleanup, so there is a funtion to position the UI. It is now called every frame instead of just when needed, which is slightly inefficient, but I operations aren't hugely expensive, just more drawing on top the masses of drawing being done anyway, and there aren't many frames where it UI wasn't being updated, as the player is never often still.